Fix compiling of sync on linux. Review URL: http://codereview.chromium.org/211019
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@26835 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 36d4ad5008ee38a8ab0ddad065e4ec46ffff5d0a

+++ chrome/browser/sync/util/character_set_converters_linux.cc
@@ -29,7 +29,7 @@ void TrimPathStringToValidCharacter(PathString* string) {
     return;
   if (0 == (string->at(string->length() - 1) & 0x080))
     return;
-  int partial_enc_bytes = 0;
+  size_t partial_enc_bytes = 0;
   for (partial_enc_bytes = 0 ; true ; ++partial_enc_bytes) {
     if (4 == partial_enc_bytes || partial_enc_bytes == string->length()) {
       // original string was broken, garbage in, garbage out.
